Remove switch in S3RequestMiddleware.virtualAddressFixup()
The two code blocks generating virtual addresses were virtually the same but were written completely differently.

The only functional difference was the default case required there to be an extra path component after the bucket name, which is not a valid requirement. This caused commands like DeleteBucket to not use virtual addressing.

At the same time I also percent encoded the path before reconstructing the URL as any percent encoding was lost when accessing url.path