Code Generator extend cyclical reference test
If a shape has a reference to itself we have to output it as a class instead of a struct. Previous the test for this would only test the immediate members of a shape. I have had to extend this to the members of member structs to catch a situation in WAFV2 where the shape `Statement` should have been output as a class.